refactor onTutorialConfig
