+++ src/actions/index.ts
@@ -1,3 +1,4 @@
 export { default as onStartup } from './onStartup'
+export { default as onTutorialConfig } from './onTutorialConfig'
 export { default as onErrorPage } from './onErrorPage'
 export { default as onTestPass } from './onTestPass'
